changed examples for static list
