+++ packages/lab/stories/static-list/static-list.qa.stories.tsx
@@ -8,87 +8,79 @@ export default {
   title: "Lab/Static List/Static List QA",
   component: StaticList,
 } as Meta<typeof StaticList>;
+
+const ListItemWithIcons = () => (
+  <StaticListItem>
+    <FlexLayout gap={1} style={{ width: "100%" }}>
+      <NotificationIcon />
+      <StackLayout gap={0.5}>
+        <Text color="inherit">Item label</Text>
+        <Text variant="secondary">Secondary label</Text>
+      </StackLayout>
+    </FlexLayout>
+  </StaticListItem>
+);
+
 const ListItem = () => (
   <StaticListItem>
-    <FlexLayout gap={1} style={{ width: "100%" }}>
-      <NotificationIcon />
-      <StackLayout gap={0.5}>
-        <Text color="inherit">Item label</Text>
-        <Text variant="secondary">Secondary label</Text>
-      </StackLayout>
-      <StackLayout direction={"row"} gap={0.5} style={{ marginLeft: "auto" }}>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-      </StackLayout>
-    </FlexLayout>
+    <StackLayout gap={0.5} style={{ width: "100%" }}>
+      <Text color="inherit">Item label</Text>
+      <Text variant="secondary">Secondary label</Text>
+    </StackLayout>
   </StaticListItem>
 );
 
-const ListItemWithoutIcons = () => (
+const ListItemWithDivider = () => (
+  <StaticListItem divider>
+    <StackLayout gap={0.5} style={{ width: "100%" }}>
+      <Text color="inherit">Item label</Text>
+      <Text variant="secondary">Secondary label</Text>
+    </StackLayout>
+  </StaticListItem>
+);
+
+const ListItemWithButtons = () => (
   <StaticListItem>
-    <FlexLayout gap={1} style={{ width: "100%" }}>
-      <StackLayout gap={0.5}>
-        <Text color="inherit">Item label</Text>
-        <Text variant="secondary">Secondary label</Text>
-      </StackLayout>
-      <StackLayout direction={"row"} gap={0.5} style={{ marginLeft: "auto" }}>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-      </StackLayout>
-    </FlexLayout>
-  </StaticListItem>
-);
-
-const NoDividerListItem = () => (
-  <StaticListItem divider={false}>
-    <FlexLayout gap={1} style={{ width: "100%" }}>
-      <NotificationIcon />
-      <StackLayout gap={0.5}>
-        <Text color="inherit">Item label</Text>
-        <Text variant="secondary">Secondary label</Text>
-      </StackLayout>
-      <StackLayout direction={"row"} gap={0.5} style={{ marginLeft: "auto" }}>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-      </StackLayout>
-    </FlexLayout>
+    <StackLayout gap={0.5} style={{ width: "100%" }}>
+      <Text color="inherit">Item label</Text>
+      <Text variant="secondary">Secondary label</Text>
+    </StackLayout>
+    <StackLayout direction={"row"} gap={0.5} style={{ marginLeft: "auto" }}>
+      <Button variant="secondary" aria-label={"icon"}>
+        <NoteIcon aria-hidden />
+      </Button>
+      <Button variant="secondary" aria-label={"icon"}>
+        <NoteIcon aria-hidden />
+      </Button>
+    </StackLayout>
   </StaticListItem>
 );
 
 export const AllExamples: StoryFn<QAContainerProps> = ({ imgSrc }) => (
-  <QAContainer cols={4} height={950} imgSrc={imgSrc} itemPadding={5}>
+  <QAContainer cols={5} height={950} imgSrc={imgSrc} itemPadding={5}>
+    <StaticList style={{ width: "320px" }}>
+      <ListItem />
+    </StaticList>
     <StaticList style={{ width: "320px" }}>
       <ListItem />
       <ListItem />
       <ListItem />
     </StaticList>
     <StaticList style={{ width: "320px" }}>
-      <ListItemWithoutIcons />
-      <ListItemWithoutIcons />
-      <ListItemWithoutIcons />
+      <ListItemWithIcons />
+      <ListItemWithIcons />
+      <ListItemWithIcons />
     </StaticList>
     <StaticList style={{ width: "320px" }}>
-      <NoDividerListItem />
-      <NoDividerListItem />
-      <NoDividerListItem />
+      <ListItemWithButtons />
+      <ListItemWithButtons />
+      <ListItemWithButtons />
+    </StaticList>
+    <StaticList style={{ width: "320px" }}>
+      <ListItem />
+      <ListItemWithDivider />
+      <ListItemWithDivider />
+      <ListItemWithDivider />
     </StaticList>
   </QAContainer>
 );
