+++ site/docs/components/static-list/examples.mdx
@@ -10,27 +10,43 @@ data:
 
 <LivePreviewControls>
 
-  <LivePreview componentName="static-list" exampleName="WithDescriptiveIcons" >
+  <LivePreview componentName="static-list" exampleName="DefaultSingle" >
 
-## With Descriptive Icons
+## Default Single
+
+Default `Static list` with a single item.
+
+  </LivePreview>
+
+  <LivePreview componentName="static-list" exampleName="DefaultMultiple" >
+
+## Default Multiple
+
+Default `Static list` group showing multiple items stacked together. 
+
+  </LivePreview>
+
+  <LivePreview componentName="static-list" exampleName="WithIcons" >
+
+## With Icons
 
 `Static list` items with descriptive icon  grouped together to organize related content. You can change the default descriptive icon with other icons based on your use-case.
 
   </LivePreview>
 
-  <LivePreview componentName="static-list" exampleName="WithoutDescriptiveIcons" >
+  <LivePreview componentName="static-list" exampleName="WithButtons" >
 
-## Without Descriptive Icons
+## With Buttons
 
-`Static list` items without descriptive icon grouped together to organize related content. The descriptive icon is optional.
+`Static List `items with interactive actions. You can change the default button based on your use-case.
 
   </LivePreview>
 
-  <LivePreview componentName="static-list" exampleName="WithoutDivider" >
+  <LivePreview componentName="static-list" exampleName="WithDividers" >
 
-## Without Divider
+## With Dividers
 
-`Static List` items grouped together without dividers
+`Static List` items grouped together with dividers
 
   </LivePreview>
