--- site/src/examples/static-list/WithDescriptiveIcons.tsx
@@ -1,38 +1,0 @@
-import { Button, FlexLayout, StackLayout, Text } from "@salt-ds/core";
-import { NoteIcon, NotificationIcon } from "@salt-ds/icons";
-import { StaticList, StaticListItem } from "@salt-ds/lab";
-import type { ReactElement } from "react";
-
-const ListItem = () => (
-  <StaticListItem>
-    <FlexLayout gap={1} style={{ width: "100%" }}>
-      <NotificationIcon />
-      <StackLayout gap={0.5}>
-        <Text color="inherit">Item label</Text>
-        <Text variant="secondary">Secondary label</Text>
-      </StackLayout>
-      <StackLayout direction={"row"} gap={0.5} style={{ marginLeft: "auto" }}>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-      </StackLayout>
-    </FlexLayout>
-  </StaticListItem>
-);
-
-export const WithDescriptiveIcons = (): ReactElement => (
-  <div style={{ width: "80%" }}>
-    <StaticList style={{ width: "320px" }}>
-      <ListItem />
-      <ListItem />
-      <ListItem />
-      <ListItem />
-      <ListItem />
-    </StaticList>
-  </div>
-);
