+++ site/src/examples/static-list/WithIcons.tsx
@@ -1,31 +1,21 @@
-import { Button, FlexLayout, StackLayout, Text } from "@salt-ds/core";
-import { NoteIcon, NotificationIcon } from "@salt-ds/icons";
+import { FlexLayout, StackLayout, Text } from "@salt-ds/core";
+import { NotificationIcon } from "@salt-ds/icons";
 import { StaticList, StaticListItem } from "@salt-ds/lab";
 import type { ReactElement } from "react";
 
 const ListItem = () => (
-  <StaticListItem divider={false}>
+  <StaticListItem>
     <FlexLayout gap={1} style={{ width: "100%" }}>
       <NotificationIcon />
       <StackLayout gap={0.5}>
         <Text color="inherit">Item label</Text>
         <Text variant="secondary">Secondary label</Text>
       </StackLayout>
-      <StackLayout direction={"row"} gap={0.5} style={{ marginLeft: "auto" }}>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-        <Button variant="secondary" aria-label={"icon"}>
-          {" "}
-          <NoteIcon aria-hidden />
-        </Button>
-      </StackLayout>
     </FlexLayout>
   </StaticListItem>
 );
 
-export const WithoutDivider = (): ReactElement => (
+export const WithIcons = (): ReactElement => (
   <div style={{ width: "80%" }}>
     <StaticList style={{ width: "320px" }}>
       <ListItem />
