+++ site/src/examples/static-list/index.ts
@@ -1,3 +1,5 @@
-export * from "./WithoutDivider";
-export * from "./WithDescriptiveIcons";
-export * from "./WithoutDescriptiveIcons";
+export * from "./WithDividers";
+export * from "./WithIcons";
+export * from "./DefaultMultiple";
+export * from "./WithButtons";
+export * from "./DefaultSingle";
